pyproject.toml
[project]
name = "resumemaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "bs4>=0.0.2",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.0",
    "svglib>=1.5.1",
    "uvicorn>=0.35.0",
]
